refactor: add useDismissToastOnRouteChange hook
