+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 import useUserProfile from "./useUserProfile";
 import useDocumentHeightResize from "./useDocumentHeightResize";
+import useDismissToastOnRouteChange from "./useDismissToastOnRouteChange";
 
-export { useUserProfile, useDocumentHeightResize };
+export { useUserProfile, useDocumentHeightResize, useDismissToastOnRouteChange };
