+++ src/routes/ApplicationRoot.tsx
@@ -1,30 +1,13 @@
-import { useAppSelector } from "@/app/hooks";
 import ScrollToTop from "@/components/ScrollToTop";
 import TopLoadingBar from "@/components/TopLoadingBar";
 import { Toaster } from "@/components/ui/toaster";
-import { useToast } from "@/components/ui/use-toast";
-import { selectNotification } from "@/features/notification/notificationSlice";
-import { useDocumentHeightResize } from "@/hooks";
-import { Fragment, useEffect } from "react";
-import { Outlet, useLocation, useMatch } from "react-router-dom";
+import { useDismissToastOnRouteChange, useDocumentHeightResize } from "@/hooks";
+import { Fragment } from "react";
+import { Outlet } from "react-router-dom";
 
 const ApplicationRoot = () => {
-  const location = useLocation();
-  const { id } = useAppSelector(selectNotification);
-  const { dismiss } = useToast();
-  const matchJobDetails = useMatch("/jobs/:jobId");
-  const matchJobListings = useMatch("/employer/job-listings");
-  const matchJobDetailsDashboard = useMatch("/employer/job-listings/:jobId");
-
   useDocumentHeightResize();
-
-  useEffect(() => {
-    if (id) {
-      if (matchJobDetails || matchJobListings || matchJobDetailsDashboard) return;
-      dismiss(id);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location.pathname, matchJobDetails, matchJobListings, matchJobDetailsDashboard]);
+  useDismissToastOnRouteChange();
 
   return (
     <Fragment>
